Moved dependency parsing helper to common
